+++ components/molecules/dog/card/__tests__/stories.js
@@ -1,0 +1,38 @@
+/**
+ * Dog - Card
+ */
+
+// React
+import React from 'react'
+
+// Component
+import { DogCard } from '../'
+import Readme from '../README.md'
+import vizla from '../__resources__/vizla.jpg'
+
+export default {
+  title: 'Molecules/Dog/Card',
+  component: DogCard,
+  parameters: {
+    readme: {
+      sidebar: Readme
+    }
+  }
+}
+
+export const defaultStory = () => (
+  <DogCard
+    breed='Cockapoo'
+    breeder='Breeder'
+    breedName='Cockapoo'
+    dogId='id'
+    gender='male'
+    image={vizla}
+    name='DogName'
+    price={1000}
+  />
+)
+
+defaultStory.story = {
+  name: 'Default'
+}
